+++ src/main/java/de/theidler/create_mobile_packages/CommonEvents.java
@@ -1,11 +1,13 @@
 package de.theidler.create_mobile_packages;
 
 import de.theidler.create_mobile_packages.blocks.bee_port.BeePortBlockEntity;
+import de.theidler.create_mobile_packages.entities.robo_entity.RoboEntity;
 import de.theidler.create_mobile_packages.robo.RoboManager;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.Level;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
+import net.neoforged.neoforge.event.entity.EntityTravelToDimensionEvent;
 import net.neoforged.neoforge.event.tick.LevelTickEvent;
 
 
@@ -28,4 +30,13 @@ public class CommonEvents {
             BeePortBlockEntity.registerCapabilities(event);
         }
     }
+
+    @SubscribeEvent
+    public static void onEntityTravelToDimension(EntityTravelToDimensionEvent event) {
+        if (event.getEntity().level().isClientSide()) return;
+        if (!(event.getEntity() instanceof RoboEntity)) return;
+
+        event.setCanceled(true);
+    }
+
 }