+++ src/main/java/de/theidler/create_mobile_packages/robo/PlayerTarget.java
@@ -21,6 +21,7 @@ import static de.theidler.create_mobile_packages.CMPHelper.doesAddressMatchPlaye
 
 public class PlayerTarget implements RoboTarget {
     private final Player player;
+    private int eta;
 
     public PlayerTarget(Player player) {
         this.player = player;
@@ -47,8 +48,7 @@ public class PlayerTarget implements RoboTarget {
         return player != null && player.isAlive();
     }
 
-    @Override
-    public void setETA(VirtualRobo robo, int eta) {
+    public void updateEtaToast(VirtualRobo robo) {
         ItemStack packageItem = robo.getItemStack();
         ItemStackHandler itemHandler = packageItem.isEmpty() ? new ItemStackHandler(9) : PackageItem.getContents(packageItem);
         List<ItemStack> items = new ArrayList<>();
@@ -59,11 +59,21 @@ public class PlayerTarget implements RoboTarget {
         PackageToast toast = new PackageToast(
                 robo.getId(),
                 Component.translatable("create_mobile_packages.toast.robo_bee_on_the_way"),
-                Component.translatable("create_mobile_packages.toast.eta", eta),
+                Component.translatable("create_mobile_packages.toast.eta", getETA()),
                 CMPItems.ROBO_BEE.asStack(),
                 items
         );
         if (player instanceof ServerPlayer serverPlayer)
             CatnipServices.NETWORK.sendToClient(serverPlayer, new ShowToastOnClientPacket(toast));
+    }
+
+    @Override
+    public int getETA() {
+        return eta;
+    }
+
+    @Override
+    public void setETA(int eta) {
+        this.eta = eta;
     }
 }