+++ src/main/java/de/theidler/create_mobile_packages/robo/VirtualRobo.java
@@ -3,6 +3,7 @@ package de.theidler.create_mobile_packages.robo;
 import com.simibubi.create.content.logistics.box.PackageItem;
 import de.theidler.create_mobile_packages.CMPHelper;
 import de.theidler.create_mobile_packages.blocks.bee_port.BeePortBlockEntity;
+import de.theidler.create_mobile_packages.blocks.bee_port.RoboRequest;
 import de.theidler.create_mobile_packages.entities.robo_entity.RoboBeeBehaviorController;
 import de.theidler.create_mobile_packages.entities.robo_entity.RoboEntity;
 import de.theidler.create_mobile_packages.index.CMPEntities;
@@ -18,8 +19,7 @@ import org.jetbrains.annotations.Nullable;
 
 import java.util.UUID;
 
-import static de.theidler.create_mobile_packages.CMPHelper.readVec3FromTag;
-import static de.theidler.create_mobile_packages.CMPHelper.writeVec3ToTag;
+import static de.theidler.create_mobile_packages.CMPHelper.*;
 
 
 public class VirtualRobo {
@@ -39,6 +39,7 @@ public class VirtualRobo {
     private float packageHeightScale;
     private int ticksSinceLastTargetCheck = 0;
     private boolean targetIsInWrongNetwork = false;
+    private RoboRequest request = null;
 
     public VirtualRobo(ServerLevel level, UUID id, ItemStack itemStack, BlockPos spawnPos, UUID logisticsNetworkId) {
         this.id = id;
@@ -105,11 +106,6 @@ public class VirtualRobo {
         // if the target is still valid and in the correct network, do nothing
         if (target != null && target.isValid() && !targetIsInWrongNetwork) return;
 
-        // check if the old target was a BeePortBlockEntity if so, then remove the reference
-        if (target != null && target.asBeePortBlockEntity() != null) {
-            target.asBeePortBlockEntity().trySetEntityOnTravel(this, false );
-        }
-
         // try finding a Player first
         target = PlayerTarget.fromAddress(serverLevel, targetAddress);
         if (target.isValid()) {
@@ -124,13 +120,6 @@ public class VirtualRobo {
             // Check if the port is in the correct network
             targetIsInWrongNetwork = !logisticsNetworkId.equals(targetBlockEntity.getLogisticsNetworkId());
         }
-        if (target.isValid() && target.asBeePortBlockEntity() != null) {
-            target.asBeePortBlockEntity().trySetEntityOnTravel(this, true );
-            return;
-        }
-        // if no valid target is found, set the target to null
-        target = null;
-        targetIsInWrongNetwork = false;
     }
 
     public float getPitch() {
@@ -174,7 +163,7 @@ public class VirtualRobo {
         updateTarget();
         if (behaviorController != null) behaviorController.tick(this);
         this.move(targetVelocity);
-        //updateNametag(); -> client side only
+        updateEta();
 
         // Spawn / despawn RoboEntity if needed
         BlockPos pos = BlockPos.containing(currentPos);
@@ -184,6 +173,14 @@ public class VirtualRobo {
             }
         } else if (entityId != null) {
             despawnEntity();
+        }
+    }
+
+    private void updateEta() {
+        if (request != null) {
+            request.setEta(calcETA(getTargetPosition(), getCurrentPos()));
+        } else if (target != null) {
+            target.setETA(calcETA(getTargetPosition(), getCurrentPos()));
         }
     }
 
@@ -283,7 +280,7 @@ public class VirtualRobo {
         return rotateToAngle((float) getSnapAngle(getAngleToTarget()) + 90);
     }
 
-    public BeePortBlockEntity getStartBeePortBlockEntity() {
+    public @Nullable BeePortBlockEntity getStartBeePortBlockEntity() {
         if (serverLevel.getBlockEntity(BlockPos.containing(currentPos)) instanceof BeePortBlockEntity bpbe) {
             return bpbe;
         } else if (serverLevel.getBlockEntity(BlockPos.containing(currentPos.subtract(0,1,0))) instanceof BeePortBlockEntity bpbe) {
@@ -296,8 +293,8 @@ public class VirtualRobo {
 
     public void setRemoved(ServerLevel level) {
         RoboManager.get(level).remove(this.getId());
-        if (target != null) {
-            target.asBeePortBlockEntity().trySetEntityOnTravel(null, false );
+        if (request != null && request.getStatus() == RoboRequest.Status.IN_PROGRESS) {
+            request.setStatus(RoboRequest.Status.PENDING);
         }
         despawnEntity();
     }
@@ -332,5 +329,20 @@ public class VirtualRobo {
 
     public void setPitch(float pitch) {
         this.pitch = pitch;
+    }
+
+    public void invalidateTarget() {
+        this.targetVelocity = Vec3.ZERO;
+        this.target = null;
+    }
+
+    public RoboRequest getRequest() {
+        return request;
+    }
+
+    public void setRequest(RoboRequest request) {
+        this.request = request;
+        this.request.setStatus(RoboRequest.Status.IN_PROGRESS);
+        this.target = new BeePortBlockEntityTarget((BeePortBlockEntity) serverLevel.getBlockEntity(request.getTargetPos()));
     }
 }